Implement first pass at crouching - #1
Crouching is controlled by a trigger / property setting. Crouching in general behaves the same as grounded, with the exception of using it as a state to track when we need to change back the Collider Height and Step Height Ratio as we exit the state.

Note that the collider height must be modified *before* the step height due to the way the mover will recalculate the sensor. Generally whatever multiplier you apply to the Height should be the same as the step height (e.g. if you reduce the collider by half, reduce step height by half) but both are supplied just in case the ratio needs to be different.