Start of Ladder functionality - Minor state refactor
Currently this supports getting onto and off ladders, albeit the off ladder is only when going "down" and kicks to the right rather than behind.

LadderZoneTrigger is the method we'll likely use for ladders, but can be implemented in other ways through ILadderInfo if desired. Might just condense into LadderZoneTrigger as requirement, but will wait until finished with ladder implementation.

Just committing now because a decent amount of work and may have gotten Win11 SSD bug and don't want to lose the work.