Minor refactor to prepare for support free climbing
LadderZone is being converted into a more generic ClimbZone trigger as the behaivor and fields are mostly the same, it'll just be a flag and the controller will consume the start/end points differently.

Additionally CurrentLadder is now a List for zones to allow for compound climb zones if desired.